Add implementation of Cursor
ForEach isn't a great option simply due to the fact that there isn't a good way
to provide an unmarshaled version of each value without getting a new value each
time. Cursor is a good alternative because it gives us the option to pass in
whatever we want for each iteration.